+++ src/AppComplianceAutomation/AppComplianceAutomation.Autorest/test/Get-AzAcatControlAssessment.Tests.ps1
@@ -1,0 +1,22 @@
+if(($null -eq $TestName) -or ($TestName -contains 'Get-AzAcatControlAssessment'))
+{
+  $loadEnvPath = Join-Path $PSScriptRoot 'loadEnv.ps1'
+  if (-Not (Test-Path -Path $loadEnvPath)) {
+      $loadEnvPath = Join-Path $PSScriptRoot '..\loadEnv.ps1'
+  }
+  . ($loadEnvPath)
+  $TestRecordingFile = Join-Path $PSScriptRoot 'Get-AzAcatControlAssessment.Recording.json'
+  $currentPath = $PSScriptRoot
+  while(-not $mockingPath) {
+      $mockingPath = Get-ChildItem -Path $currentPath -Recurse -Include 'HttpPipelineMocking.ps1' -File
+      $currentPath = Split-Path -Path $currentPath -Parent
+  }
+  . ($mockingPath | Select-Object -First 1).FullName
+}
+
+Describe 'Get-AzAcatControlAssessment' {
+    It 'List' {
+        $categories = Get-AzAcatControlAssessment -ReportName $env.GeneratedReportName
+        $categories.Count | Should -BeGreaterOrEqual 1
+    }
+}