+++ src/Storage/Storage.Management/StorageAccount/NewAzureStorageAccount.cs
@@ -496,6 +496,7 @@ namespace Microsoft.Azure.Commands.Management.Storage
         [ValidateSet(StorageModels.MinimumTlsVersion.TLS10,
             StorageModels.MinimumTlsVersion.TLS11,
             StorageModels.MinimumTlsVersion.TLS12,
+            StorageModels.MinimumTlsVersion.TLS13,
             IgnoreCase = true)]
         public string MinimumTlsVersion
         {
@@ -876,7 +877,7 @@ namespace Microsoft.Azure.Commands.Management.Storage
             }
             if (sasExpirationPeriod != null)
             {
-                createParameters.SasPolicy = new SasPolicy(sasExpirationPeriod.Value.ToString(@"d\.hh\:mm\:ss"));
+                createParameters.SasPolicy = new SasPolicy(sasExpirationPeriod.Value.ToString(@"d\.hh\:mm\:ss"), "Log");
             }
             if (keyExpirationPeriodInDay != null)
             {